requirements.txt
amazon-braket-default-simulator==1.3.0
amazon-braket-ocean-plugin==1.0.6
amazon-braket-schemas==1.4.1
amazon-braket-sdk==1.9.5.post0
attrs==21.2.0
backoff==1.11.1
boltons==21.0.0
boto3==1.20.8
botocore==1.23.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coverage==5.5
cplex==20.1.0.1
cryptography==35.0.0
cycler==0.11.0
decorator==4.4.2
dill==0.3.4
dimod==0.10.8
diskcache==5.2.1
dlx==1.0.4
docplex==2.22.213
dwave-cloud-client==0.9.2
dwave-greedy==0.2.1
dwave-inspector==0.2.7
dwave-neal==0.5.8
dwave-networkx==0.8.10
dwave-preprocessing==0.3.1.post0
dwave-qiskit-plugin==0.1.0
dwave-system==1.10.0
dwave-tabu==0.4.2
fastdtw==0.3.4
fasteners==0.16.3
fastjsonschema==2.15.1
Flask==2.0.2
fonttools==4.28.1
gurobipy==9.5.0
h5py==3.2.1
homebase==1.0.1
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonref==0.2
jsonschema==4.2.1
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
minorminer==0.2.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.11.0
mpmath==1.2.1
multitasking==0.0.10
nest-asyncio==1.5.1
networkx==2.5.1
ntlm-auth==1.5.0
numpy==1.21.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pip==21.2.4
plucky==0.4.3
ply==3.11
psutil==5.8.0
pycparser==2.21
pydantic==1.8.2
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-constraint==1.4.0
python-dateutil==2.8.2
pytz==2021.3
qiskit-aqua==0.9.5
qiskit-ibmq-provider==0.18.0
qiskit-ignis==0.6.0
qiskit-optimization==0.2.3
qiskit-terra==0.18.3
Quandl==3.7.0
requests==2.26.0
requests-ntlm==1.1.0
retworkx==0.10.2
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.2
setuptools==58.0.4
setuptools-scm==6.3.2
six==1.16.0
symengine==0.8.1
sympy==1.9
threadpoolctl==3.0.0
tomli==1.2.2
tweedledum==1.1.1
typing_extensions==4.0.0
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
wheel==0.37.0
yfinance==0.1.66
zipp==3.6.0
